fix bugs

+++ backend/controlers/post.js
@@ -2,13 +2,11 @@ const db = require("../db");
 const passport = require("passport");
 const multer = require("multer");
 const { query, body, validationResult } = require("express-validator");
-const path = require("node:path");
+const path = require("path");
 const { queryDb } = require("../utils");
-const { rmSync } = require("node:fs");
+const fs = require("@cyclic.sh/s3fs");
+
 const storage = multer.diskStorage({
-  destination: (req, file, cb) => {
-    cb(null, "uploads/");
-  },
   filename: (req, file, cb) => {
     cb(null, Date.now() + path.extname(file.originalname));
   },
@@ -32,28 +30,49 @@ exports.create_post = [
   upload.single("image"),
   async (req, res, next) => {
     try {
+
       const errors = validationResult(req);
       if (!errors.isEmpty()) {
         res.status(400).json({
           errors,
         });
       } else {
+        
         const createdDate = new Date();
+        const year = createdDate.getFullYear();
+        const month = `${createdDate.getMonth() + 1}`.padStart(2, "0");
+        const day = `${createdDate.getDate()}`.padStart(2, "0");
+
+        // get the hours, minutes, and seconds as separate strings
+        const hours = `${createdDate.getHours()}`.padStart(2, "0");
+        const minutes = `${createdDate.getMinutes()}`.padStart(2, "0");
+        const seconds = `${createdDate.getSeconds()}`.padStart(2, "0");
+
+        // combine the createdDate and time strings into a single string in the desired format
+        const dateTimeString = `${year}-${month}-${day} ${hours}:${minutes}:${seconds}`;
         const postdetails = [
           req.user,
           req.body.title,
           req.body.content,
           req.body.price,
-          createdDate,
+          dateTimeString,
         ];
         const imageRegEx = /\.(gif|jpe?g|jfif|tiff?|png|webp|bmp)$/i;
+        let query =
+          "INSERT INTO posts (authorid, title, content, price, createddate";
         if (req.file && imageRegEx.test(req.file.filename)) {
           postdetails.push(req.file.path.replace(/\\/g, "/"));
+          query += ",image) VALUES (?,?,?,?,?,?)";
+          fs.writeFile(req.file.filename, req.file.buffer, function(err) {
+            if (err) {
+              console.error(err);
+              return res.status(500).send('Error uploading file');
+            }
+          });
+        } else {
+          query += ") VALUES (?,?,?,?,?)";
         }
-        await queryDb(
-          "INSERT INTO posts (authorid, title, content, price, createddate, image) VALUES(?,?,?,?,?,?)",
-          postdetails
-        );
+        await queryDb(query, postdetails);
 
         let categories = [];
         if (req.body.categories) {
@@ -63,27 +82,14 @@ exports.create_post = [
                 "SELECT id from categories where name = ?",
                 cat
               );
-              return result[0].id;
+              const r = result[0];
+              return r.id;
             })
           );
-          const year = createdDate.getFullYear();
-          const month = `${createdDate.getMonth() + 1}`.padStart(2, "0");
-          const day = `${createdDate.getDate()}`.padStart(2, "0");
-
-          // get the hours, minutes, and seconds as separate strings
-          const hours = `${createdDate.getHours()}`.padStart(2, "0");
-          const minutes = `${createdDate.getMinutes()}`.padStart(2, "0");
-          const seconds = `${createdDate.getSeconds()}`.padStart(2, "0");
-
-          // combine the createdDate and time strings into a single string in the desired format
-          const dateTimeString = `${year}-${month}-${day} ${hours}:${minutes}:${seconds}`;
-
           const post = await queryDb(
             "SELECT * FROM posts WHERE title = ? AND authorid = ? AND createddate LIKE ?",
             [req.body.title, req.user, dateTimeString]
           );
-          console.log(categories);
-
           categories.forEach(async (e) => {
             await queryDb(
               "INSERT INTO post_categories (postid, categoryid) VALUES (?, ?)",
@@ -91,7 +97,7 @@ exports.create_post = [
             );
           });
         }
-        console.log(categories);
+
         res.status(200).json({
           msg: "تم انشاء منشور بنجاح",
         });
@@ -154,6 +160,12 @@ exports.edit_post = [
     if (req.file && imageRegEx.test(req.file.filename)) {
       postdetails.push(req.file.path.replace(/\\/g, "/"));
       query += ", image = ?";
+      fs.writeFile(req.file.filename, req.file.buffer, function(err) {
+        if (err) {
+          console.error(err);
+          return res.status(500).send('Error uploading file');
+        }
+      });
     }
     postdetails.push(new Date());
     postdetails.push(req.body.postid);
@@ -187,35 +199,42 @@ exports.add_like = [
   }),
   async (req, res, next) => {
     try {
-      const result = await queryDb(
-        "SELECT * FROM post_likes WHERE user_id = ? AND post_id = ?",
-        [req.user, +req.body.postid]
+      const post = await queryDb(
+        "SELECT  * FROM posts WHERE id = ? AND authorid = ?",
+        [req.body.postid, req.user]
       );
-      console.log(result);
-      if (result.length > 0) {
-        await queryDb(
-          "DELETE FROM post_likes WHERE user_id = ? AND post_id = ?",
+      if (post.length > 0) {
+        const result = await queryDb(
+          "SELECT * FROM post_likes WHERE user_id = ? AND post_id = ?",
           [req.user, +req.body.postid]
         );
-        await queryDb(
-          "UPDATE posts SET likes_count = likes_count -1 WHERE authorid = ? AND id = ?",
-          [req.user, +req.body.postid]
-        );
-        return res.status(200).json({
-          msg: "تم الغاء الاعجاب بالمنشور بنجاح",
-        });
+        if (result.length > 0) {
+          await queryDb(
+            "DELETE FROM post_likes WHERE user_id = ? AND post_id = ?",
+            [req.user, +req.body.postid]
+          );
+          await queryDb(
+            "UPDATE posts SET likes_count = likes_count -1 WHERE authorid = ? AND id = ?",
+            [req.user, +req.body.postid]
+          );
+          return res.status(200).json({
+            msg: "تم الغاء الاعجاب بالمنشور بنجاح",
+          });
+        } else {
+          await queryDb(
+            "INSERT INTO post_likes (user_id, post_id) VALUES(?,?)",
+            [req.user, +req.body.postid]
+          );
+          await queryDb(
+            "UPDATE posts SET likes_count = likes_count + 1 WHERE authorid = ? AND id = ?",
+            [req.user, +req.body.postid]
+          );
+          return res.status(200).json({
+            msg: "تم الاعجاب بالمنشور بنجاح",
+          });
+        }
       } else {
-        await queryDb("INSERT INTO post_likes (user_id, post_id) VALUES(?,?)", [
-          req.user,
-          +req.body.postid,
-        ]);
-        await queryDb(
-          "UPDATE posts SET likes_count = likes_count + 1 WHERE authorid = ? AND id = ?",
-          [req.user, +req.body.postid]
-        );
-        return res.status(200).json({
-          msg: "تم الاعجاب بالمنشور بنجاح",
-        });
+        res.status(404).json("لم يتم العثور على المنشور المطلوب");
       }
     } catch (err) {
       console.error(err);
@@ -240,7 +259,6 @@ exports.get_posts = [
       ORDER BY p.createddate DESC
       LIMIT 10 OFFSET ${req.query.page > 0 ? req.query.page * 10 : 0};`;
       const posts = await queryDb(query);
-      console.log(req.params.page);
       if (posts.length > 0) {
         return res.status(200).json({
           posts: posts,
@@ -275,9 +293,9 @@ exports.get_post = [
       p.likes_count, p.createddate, CONCAT(u.firstname, ' ', u.lastname) AS authorname, 
       u.phonenumber, u.photo FROM posts AS p
       INNER JOIN users AS  u ON p.authorid = u.id
-      WHERE p.id = 56;`;
+      WHERE p.id = ?;`;
 
-      const post = await queryDb(query);
+      const post = await queryDb(query, req.query.postid);
       if (post.length > 0) {
         res.status(200).json({
           post: post,
@@ -298,29 +316,33 @@ exports.delete_post = [
   passport.authenticate("jwt", { session: false }),
   body("postid").escape(),
   async (req, res, next) => {
-    if (req.body.postid) {
-      const post = await queryDb(
-        "SELECT * FROM posts WHERE id = ? AND authorid = ?",
-        [req.body.postid, req.user]
-      );
-      if (post.length > 0) {
-        await queryDb("DELETE FROM posts WHERE id = ? AND authorid = ?", [
-          req.body.postid,
-          req.user,
-        ]);
-        return res.status(200).json({
-          msg: 'تم حذف المنشور بنجاح',
+    try {
+      if (req.body.postid) {
+        const post = await queryDb(
+          "SELECT * FROM posts WHERE id = ? AND authorid = ?",
+          [req.body.postid, req.user]
+        );
+        if (post.length > 0) {
+          await queryDb("DELETE FROM posts WHERE id = ? AND authorid = ?", [
+            req.body.postid,
+            req.user,
+          ]);
+          return res.status(200).json({
+            msg: "تم حذف المنشور بنجاح",
+          });
+        } else {
+          return res.status(404).json({
+            msg: "المنشور غير موجود ",
+          });
+        }
+      } else {
+        return res.status(400).json({
+          msg: "معرف المنشور غير موجود",
         });
       }
-      else { 
-        return res.status(404).json({
-          msg: 'المنشور غير موجود ',
-        })
-      }
-    } else {
-      return res.status(400).json({
-        msg: "معرف المنشور غير موجود",
-      });
+    } catch (err) {
+      console.log(err);
+      next(err);
     }
   },
 ];
